lesson9_blog/n77qsfq/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
psycopg2==2.9.12
sqlparse==0.5.5
tzdata==2026.2
